requirements.txt
arrow==1.2.3
asgiref==3.6.0
beautifulsoup4==4.12.0
crispy-bootstrap5==0.7
dateparser==1.1.8
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-bower==5.2.0
django-clear-cache==0.3
django-crispy-forms==2.0
django-jquery==3.1.0
django-jquery-ui==1.11.4.1
django-scheduler==0.10.1
django-widget-tweaks==1.4.12
environs==9.5.0
fanstatic==1.3
icalendar==5.0.4
js.fullcalendar==2.9.1
js.jquery==3.6.1
js.momentjs==2.13.1
marshmallow==3.19.0
moment==0.12.1
packaging==23.0
Pillow==9.4.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
times==0.7
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
WebOb==1.8.7
